data.h5tools: Using 64 bit ints for buffer size to avoid int overflow.

+++ src/snomtools/data/h5tools.py
@@ -406,8 +406,8 @@ def buffer_needed(shape=None, access=None, chunks=None, data=None, dtype=None, s
             for c in range(n_chunks):
                 if c in chunk_alignment[access[dim]]:
                     chunks_needed[dim] += 1
-    chunks_needed = np.prod(chunks_needed)  # Total chunks is product of chunks of each dimension.
-    elements_needed = chunks_needed * np.prod(chunks)  # Elements per chunk is product of chunk size.
+    chunks_needed = np.prod(chunks_needed, dtype=np.uint64)  # Total chunks is product of chunks of each dimension.
+    elements_needed = chunks_needed * np.prod(chunks, dtype=np.uint64)  # Elements per chunk is product of chunk size.
     return elements_needed * np.dtype(dtype).itemsize + safety_margin
